feat(auth): implement AuthController with Google authentication routes and remove old type definition
