+++ apps/server/src/auth/auth.controller.ts
@@ -1,0 +1,34 @@
+import { Controller, Get, Post, Req, Res, UseGuards } from '@nestjs/common';
+import { AuthService } from './auth.service';
+import { GoogleAuthGuard } from './guards/google-auth/google-auth.guard';
+import { RefreshAuthGuard } from './guards/refresh-auth/refresh-auth.guard';
+
+@Controller('auth')
+export class AuthController {
+  constructor(private readonly authService: AuthService) {}
+
+  @UseGuards(GoogleAuthGuard)
+  @Get('google/login')
+  googleLogin() {}
+
+  @UseGuards(GoogleAuthGuard)
+  @Get('google/callback')
+  async googleCallback(@Req() req, @Res() res) {
+    const response = await this.authService.login(req.user.id);
+    res.redirect(
+      `${process.env.CLIENT_URL}/auth/callback?token=${response.accessToken}&refreshToken=${response.refreshToken}`,
+    );
+  }
+
+  @Post('login')
+  async login(@Req() req) {
+    const token = await this.authService.login(req.user.id);
+    return { id: req.user.id, token };
+  }
+
+  @UseGuards(RefreshAuthGuard)
+  @Post('refresh')
+  refreshToken(@Req() req) {
+    return this.authService.refreshToken(req.user.id);
+  }
+}
