--- apps/server/src/auth/types/auth.controller.ts
@@ -1,15 +1,0 @@
-import { Controller, Get, UseGuards } from '@nestjs/common';
-import { GoogleAuthGuard } from '../guards/google-auth/google-auth.guard';
-
-@Controller('auth')
-export class AuthController {
-  constructor() {}
-
-  @UseGuards(GoogleAuthGuard)
-  @Get('google/login')
-  googleLogin() {}
-
-  @UseGuards(GoogleAuthGuard)
-  @Get('google/callback')
-  googleCallback() {}
-}
